+++ src/main/java/net/silentchaos512/scalinghealth/client/ClientHandler.java
@@ -21,6 +21,7 @@ public final class ClientHandler {
     public static float worldDifficulty;
     public static float areaDifficulty;
     public static int regenTimer;
+    public static int locationMultiPercent;
     // Infrequent updates (join server/world, travel to new dimension)
     public static AreaDifficultyMode areaMode;
     public static float maxDifficultyValue;
@@ -28,14 +29,16 @@ public final class ClientHandler {
     private ClientHandler() {}
 
     public static void onMessage(ClientSyncMessage msg, Supplier<NetworkEvent.Context> ctx) {
-        //noinspection OverlyLongLambda
-        ctx.get().enqueueWork(() -> {
-            playerDifficulty = msg.playerDifficulty;
-            worldDifficulty = msg.worldDifficulty;
-            areaDifficulty = msg.areaDifficulty;
-            regenTimer = msg.regenTimer;
-        });
+        ctx.get().enqueueWork(() -> handleSyncMessage(msg));
         ctx.get().setPacketHandled(true);
+    }
+
+    private static void handleSyncMessage(ClientSyncMessage msg) {
+        playerDifficulty = msg.playerDifficulty;
+        worldDifficulty = msg.worldDifficulty;
+        areaDifficulty = msg.areaDifficulty;
+        regenTimer = msg.regenTimer;
+        locationMultiPercent = msg.locationMultiPercent;
     }
 
     public static void onLoginMessage(ClientLoginMessage msg, Supplier<NetworkEvent.Context> ctx) {