+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
 ## [Unreleased]
 ### Added
+- Location multiplier configs. Allows difficulty to be increased/decreased in different dimensions and biomes.
+
+## [2.0.3] - 2019-04-12
+### Added
 - Blight potion configs (found in dimension config files)
+- Player bonus regen configs
 ### Fixed
 - Blight fires